Add files via upload
This is a simple Tic-Tac-Toe game project I created as a beginner in web development. The project was developed using HTML, CSS, and JavaScript.I learned a lot during this project, and I took help from various YouTube tutorials to understand the concepts and implementation. This project helped me gain practical experience in web development, and I'm excited to continue improving my skills.

Feel free to check out the code and play the game!